+++ src/events/index.ts
@@ -4,7 +4,6 @@ import guildMemberAdd from './guildMemberAdd'
 import guildMemberRemove from './guildMemberRemove'
 import guildMemberUpdate from './guildMemberUpdate'
 import interactionCreate from './interactionCreate'
-import messageCreate from './messageCreate'
 import preventEmojiSpam from './preventEmojiSpam'
 import ready from './ready'
 import stickyMessageHandler from './stickyMessageHandler'
@@ -16,7 +15,6 @@ export default [
   guildMemberRemove,
   guildMemberUpdate,
   interactionCreate,
-  messageCreate,
   preventEmojiSpam,
   stickyMessageHandler,
   stickyMessageCreate,
