--- src/events/messageCreate.ts
@@ -1,17 +1,0 @@
-import { Events } from 'discord.js'
-
-import { checkName } from '@/features/nameChecker'
-import { defineEventHandler } from '@/types/defineEventHandler'
-
-export default defineEventHandler({
-  eventName: Events.MessageCreate,
-  once: true,
-  execute: async (botContext, message) => {
-    if (message.member) {
-      await checkName(
-        message.member,
-        botContext.runtimeConfiguration.data.nameChecker,
-      )
-    }
-  },
-})
