+++ src/features/index.ts
@@ -1,0 +1,5 @@
+import { Plugin } from '@/types/Plugin'
+
+import nameChecker from './nameChecker'
+
+export default [nameChecker] as Plugin[]
